requirements.txt
streamlit>=1.28.0
pypdf>=3.17.0
reportlab>=4.0.0
Pillow>=10.0.0
python-bidi>=0.6.0
